rename: component 패키지에 컴포넌트 분리
