+++ core/calendar/src/main/java/com/cheeruplife/core/calendar/components/ExpandScheduleCalendar.kt
@@ -1,9 +1,10 @@
-package com.cheeruplife.core.calendar
+package com.cheeruplife.core.calendar.components
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -12,31 +13,38 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.cheeruplife.core.calendar.model.Calendar
+import com.cheeruplife.core.calendar.model.Week
+import com.cheeruplife.core.designsystem.theme.LifeGray200
 import com.cheeruplife.core.model.Date
 import com.cheeruplife.core.model.Schedule
 
 @Composable
-fun LifeSmallScheduleCalendar(
-    weekIndex: Int,
+fun LifeExpandScheduleCalendar(
     selectDate: Date?,
-    week: List<Date>,
+    days: List<Week>,
     schedule: Map<Date, List<Schedule>>,
     onDateSelect: (Date, Int) -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    val selectEvent = remember {
-        { date: Date -> onDateSelect(date, weekIndex) }
-    }
     Column (
         modifier = modifier.fillMaxSize(),
     ) {
-        WeekItem(
-            modifier = modifier.weight(1f),
-            week = week,
-            schedule = schedule,
-            selectDate = selectDate,
-            onDateSelect = selectEvent,
-        )
+        days.forEachIndexed { weekIndex, days ->
+            val selectEvent = remember {
+                { date: Date -> onDateSelect(date, weekIndex) }
+            }
+            Column(
+                modifier = modifier.weight(1f),
+            ) {
+                HorizontalDivider(color = LifeGray200)
+                WeekItem(
+                    week = days,
+                    schedule = schedule,
+                    selectDate = selectDate,
+                    onDateSelect = selectEvent,
+                )
+            }
+        }
     }
 }
 
@@ -49,10 +57,10 @@ private fun WeekItem(
     modifier: Modifier = Modifier,
 ) {
     Row(
-        modifier = modifier.fillMaxSize(),
+        modifier = modifier.fillMaxWidth(),
     ) {
         week.forEach { date ->
-            LifeNormalCalendarItem (
+            LifeExpandCalendarItem(
                 modifier = modifier.weight(1f),
                 date = date,
                 scheduleList = schedule[date] ?: emptyList(),
@@ -63,54 +71,48 @@ private fun WeekItem(
     }
 }
 
-@Preview(name = "SmallScheduleCalendar")
+@Preview(name = "ExpandScheduleCalendar")
 @Composable
-private fun PreviewSmallScheduleCalendar() {
-    val days = remember { Calendar.create(2025, 5).days[0] }
+private fun PreviewExpandCalendar() {
+    val days = remember { Calendar.create(2025, 5).days }
     val schedule = mapOf(
-        days[0] to listOf(
+        days[0][0] to listOf(
             Schedule(
-                date = days[0],
+                date = days[0][0],
                 content = "기차표 예매해야함",
-                isHolyDay = true,
             ),
             Schedule(
-                date = days[0],
+                date = days[0][0],
                 content = "기차표 예매해야함",
             ),
         ),
-        days[1] to listOf(
+        days[1][3] to listOf(
             Schedule(
-                date = days[1],
+                date = days[1][3],
                 content = "기차표 예매해야함",
-                isCompleted = true,
+                isCompleted = true
             ),
             Schedule(
-                date = days[1],
+                date = days[1][3],
                 content = "기차표 예매해야함",
             ),
             Schedule(
-                date = days[1],
+                date = days[1][3],
                 content = "기차표 예매해야함",
             ),
         ),
     )
     var selectDate by remember {
-        val today = days.find { it.isToday }
+        val today = days.flatten().find { it.isToday }
         mutableStateOf(today)
     }
-    val selectEvent = { date: Date, index: Int ->
+    val selectEvent = { date: Date, week: Int ->
         selectDate = date
     }
-    Box(
-        modifier = Modifier.fillMaxSize(),
-    ) {
-        LifeSmallScheduleCalendar(
-            weekIndex = 1,
-            selectDate = selectDate,
-            week = days,
-            schedule = schedule,
-            onDateSelect = selectEvent,
-        )
-    }
+    LifeExpandScheduleCalendar(
+        selectDate = selectDate,
+        days = days,
+        schedule = schedule,
+        onDateSelect = selectEvent,
+    )
 }
