+++ core/calendar/src/main/java/com/cheeruplife/core/calendar/components/NormalCalendarItem.kt
@@ -1,4 +1,4 @@
-package com.cheeruplife.core.calendar
+package com.cheeruplife.core.calendar.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -17,7 +17,6 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.cheeruplife.core.designsystem.common.Dimens
