+++ core/calendar/src/main/java/com/cheeruplife/core/calendar/components/SmallScheduleCalendar.kt
@@ -1,4 +1,4 @@
-package com.cheeruplife.core.calendar
+package com.cheeruplife.core.calendar.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -12,33 +12,31 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.cheeruplife.core.calendar.model.Calendar
-import com.cheeruplife.core.calendar.model.Week
 import com.cheeruplife.core.model.Date
 import com.cheeruplife.core.model.Schedule
 
 @Composable
-fun LifeNormalScheduleCalendar(
+fun LifeSmallScheduleCalendar(
+    weekIndex: Int,
     selectDate: Date?,
-    days: List<Week>,
+    week: List<Date>,
     schedule: Map<Date, List<Schedule>>,
     onDateSelect: (Date, Int) -> Unit,
     modifier: Modifier = Modifier,
 ) {
+    val selectEvent = remember {
+        { date: Date -> onDateSelect(date, weekIndex) }
+    }
     Column (
         modifier = modifier.fillMaxSize(),
     ) {
-        days.forEachIndexed { weekIndex, days ->
-            val selectEvent = remember {
-                { date: Date -> onDateSelect(date, weekIndex) }
-            }
-            WeekItem(
-                modifier = modifier.weight(1f),
-                week = days,
-                schedule = schedule,
-                selectDate = selectDate,
-                onDateSelect = selectEvent,
-            )
-        }
+        WeekItem(
+            modifier = modifier.weight(1f),
+            week = week,
+            schedule = schedule,
+            selectDate = selectDate,
+            onDateSelect = selectEvent,
+        )
     }
 }
 
@@ -65,40 +63,40 @@ private fun WeekItem(
     }
 }
 
-@Preview(name = "NormalScheduleCalendar")
+@Preview(name = "SmallScheduleCalendar")
 @Composable
-private fun PreviewNormalScheduleCalendar() {
-    val days = remember { Calendar.create(2025, 5).days }
+private fun PreviewSmallScheduleCalendar() {
+    val days = remember { Calendar.create(2025, 5).days[0] }
     val schedule = mapOf(
-        days[0][0] to listOf(
+        days[0] to listOf(
             Schedule(
-                date = days[0][0],
+                date = days[0],
                 content = "기차표 예매해야함",
                 isHolyDay = true,
             ),
             Schedule(
-                date = days[0][0],
+                date = days[0],
                 content = "기차표 예매해야함",
             ),
         ),
-        days[1][3] to listOf(
+        days[1] to listOf(
             Schedule(
-                date = days[1][3],
+                date = days[1],
                 content = "기차표 예매해야함",
-                isCompleted = true
+                isCompleted = true,
             ),
             Schedule(
-                date = days[1][3],
+                date = days[1],
                 content = "기차표 예매해야함",
             ),
             Schedule(
-                date = days[1][3],
+                date = days[1],
                 content = "기차표 예매해야함",
             ),
         ),
     )
     var selectDate by remember {
-        val today = days.flatten().find { it.isToday }
+        val today = days.find { it.isToday }
         mutableStateOf(today)
     }
     val selectEvent = { date: Date, index: Int ->
@@ -107,9 +105,10 @@ private fun PreviewNormalScheduleCalendar() {
     Box(
         modifier = Modifier.fillMaxSize(),
     ) {
-        LifeNormalScheduleCalendar(
-            selectDate  = selectDate,
-            days = days,
+        LifeSmallScheduleCalendar(
+            weekIndex = 1,
+            selectDate = selectDate,
+            week = days,
             schedule = schedule,
             onDateSelect = selectEvent,
         )
